requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-qdrant
langchain-classic
qdrant-client
pypdf
python-dotenv